+++ components/brave_wallet_ui/components/desktop/popup-modals/accounts-modal/style.ts
