+++ components/brave_wallet_ui/components/desktop/popup-modals/add-account-modal/add-imported-account-modal.tsx
@@ -42,7 +42,7 @@ import {
 import {
   WarningText,
   WarningWrapper
-} from '../account-settings-modal/style'
+} from '../accounts-modal/style'
 
 interface Params {
   accountTypeName: string
